requirements.txt
cairocffi==0.8.1
CairoSVG==2.1.3
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.0
django-allauth==0.36.0
django-autocomplete-light==3.2.10
django-crispy-forms==1.7.0
django-weasyprint==0.5.2
djangorestframework==3.8.2
html5lib==1.0.1
httplib2==0.11.3
idna==2.7
oauthlib==2.1.0
pdfrw==0.4
Pillow==5.1.0
pycparser==2.18
PyPDF2==1.26.0
Pyphen==0.9.4
python3-openid==3.1.0
pytz==2017.3
reportlab==3.4.0
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
tinycss2==0.6.1
urllib3==1.23
WeasyPrint==0.42.3
webencodings==0.5.1
